executor: migrate test-infra to testify for testClusterTableSuite (#33886)
close pingcap/tidb#33436
